Move DDL Command Lock Acquisition (#7199)
Switched DDL commands to use more specific executor locks (read/write as necessary).
Added more concurrency tests for DDL commands, including a testing interface that can hopefully be reused for similar tests.
Cleaned up lock acquisition code.
Made error messaging for missing table exceptions more consistent.
Updated some DDL codepaths to catch and ignore TableNotFound exceptions.

+++ Catalog/SysCatalog.cpp
@@ -1115,7 +1115,7 @@ void SysCatalog::dropUserUnchecked(const std::string& name, const UserMetadata& 
   sqliteConnector_->query("END TRANSACTION");
 }
 
-void SysCatalog::dropUser(const string& name) {
+void SysCatalog::dropUser(const string& name, bool if_exists) {
   sys_write_lock write_lock(this);
   sys_sqlite_lock sqlite_lock(this);
 
@@ -1123,7 +1123,11 @@ void SysCatalog::dropUser(const string& name) {
 
   UserMetadata user;
   if (!getMetadataForUser(name, user)) {
-    throw runtime_error("Cannot drop user. User " + loggable + "does not exist.");
+    if (if_exists) {
+      return;
+    } else {
+      throw runtime_error("Cannot drop user. User " + loggable + "does not exist.");
+    }
   }
 
   if (user.userId == shared::kRootUserId) {
